+++ src/ringhopper-primitives/src/primitive/array/test.rs
@@ -4,8 +4,8 @@ use super::*;
 #[test]
 fn parse_bounds() {
     let array_of_bounds: &[u8] = &[0xBF, 0x80, 0x00, 0x00, 0x3F, 0x80, 0x00, 0x00];
-    let b = Bounds::<f64>::read_from_tag_file(&array_of_bounds, 0, 8, &mut 8).expect("should work");
-    assert_eq!(Bounds { lower: -1.0f64, upper: 1.0f64 }, b);
+    let b = Bounds::<f32>::read_from_tag_file(&array_of_bounds, 0, 8, &mut 8).expect("should work");
+    assert_eq!(Bounds { lower: -1.0f32, upper: 1.0f32 }, b);
 }
 
 #[test]
@@ -14,7 +14,7 @@ fn parse_array() {
         0x3F, 0x80, 0x00, 0x00, 0x40, 0x00, 0x00, 0x00, 0x40, 0x40, 0x00, 0x00
     ];
 
-    let floats = <[f64; 3]>::read::<BigEndian>(&array_of_references, 0, 12).expect("should be OK");
+    let floats = <[f32; 3]>::read::<BigEndian>(&array_of_references, 0, 12).expect("should be OK");
     assert_eq!(floats, [1.0, 2.0, 3.0]);
 
     let mut result = [0u8; 12];
